+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Fix signature matching for externals when abstract alias hides function arity. https://github.com/rescript-lang/rescript/pull/8045
 - Fix arity detection for arrows returning nested generics. https://github.com/rescript-lang/rescript/pull/8064
 - Fix error handling when rescript.json parsing fails and improve error message. https://github.com/rescript-lang/rescript/pull/8067
+- Fix invalid JSX being generated for empty fragments. https://github.com/rescript-lang/rescript/pull/8077
 
 #### :memo: Documentation